+++ Sources/Sentry/SentryHub.m
@@ -14,6 +14,7 @@
 #import "SentrySerialization.h"
 #import "SentryTracer.h"
 #import "SentryTracesSampler.h"
+#import "SentryTransaction.h"
 #import "SentryTransactionContext.h"
 
 @interface
@@ -228,6 +229,13 @@ SentryHub ()
     [client captureCrashEvent:event withScope:self.scope];
 }
 
+- (SentryId *)captureTransaction:(SentryTransaction *)transaction withScope:(SentryScope *)scope
+{
+    if (transaction.trace.context.sampled != kSentrySampleDecisionYes)
+        return SentryId.empty;
+    return [self captureEvent:transaction withScope:scope];
+}
+
 - (SentryId *)captureEvent:(SentryEvent *)event
 {
     return [self captureEvent:event withScope:[[SentryScope alloc] init]];
@@ -284,6 +292,17 @@ SentryHub ()
                                   bindToScope:(BOOL)bindToScope
                         customSamplingContext:(NSDictionary<NSString *, id> *)customSamplingContext
 {
+    return [self startTransactionWithContext:transactionContext
+                                 bindToScope:bindToScope
+                             waitForChildren:NO
+                       customSamplingContext:customSamplingContext];
+}
+
+- (id<SentrySpan>)startTransactionWithContext:(SentryTransactionContext *)transactionContext
+                                  bindToScope:(BOOL)bindToScope
+                              waitForChildren:(BOOL)waitForChildren
+                        customSamplingContext:(NSDictionary<NSString *, id> *)customSamplingContext
+{
     SentrySamplingContext *samplingContext =
         [[SentrySamplingContext alloc] initWithTransactionContext:transactionContext
                                             customSamplingContext:customSamplingContext];
@@ -291,7 +310,8 @@ SentryHub ()
     transactionContext.sampled = [_sampler sample:samplingContext];
 
     id<SentrySpan> tracer = [[SentryTracer alloc] initWithTransactionContext:transactionContext
-                                                                         hub:self];
+                                                                         hub:self
+                                                             waitForChildren:waitForChildren];
     if (bindToScope)
         _scope.span = tracer;