+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix: SpanProtocol add setData for Swift (#1305)
+- fix: SentryHub not checking spanContext sampled value (#1318)
 
 ## 7.2.7